Mixed-channel drips, async bounce handling for SES inbound, OOO/auto-reply
detection, time-series analytics, tag_assignments.entity_id schema fix,
and step-level connection resolvers (resolve_*_for_step) so mixed-channel
sends can resolve both email and SMS senders.

255 specs total green. No platform-level fallback paths introduced anywhere.

+++ app/services/audiences/filter_compiler.rb
@@ -166,11 +166,10 @@ module Audiences
 
       table = scope.model.table_name
       entity_type = scope.model.name
-      # tag_assignments.entity_id is varchar — explicit cast required for join.
       tag_join_sql = <<~SQL.squish
         INNER JOIN tag_assignments
           ON tag_assignments.entity_type = '#{entity_type}'
-         AND tag_assignments.entity_id::bigint = #{table}.id
+         AND tag_assignments.entity_id = #{table}.id
         INNER JOIN tags
           ON tags.id = tag_assignments.tag_id
       SQL
